+++ CHANGELOG.rst
@@ -1,6 +1,12 @@
 Changelog
 =========
 
+0.6.0 (2022-01-15)
+------------------
+
+- Use ``-m`` option instead of argument to scan a module.
+- Document the types of slot errors.
+
 0.5.3 (2022-01-14)
 ------------------